refactor(onboarding): update folder structure
