+++ src/renderer/src/features/onboarding/components/onboarding-dialog.tsx
