+++ src/renderer/src/features/onboarding/components/onboarding-modal.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 import {
 	OnboardingStep,
 	useOnboardingStore,
-} from "../../store/onboarding-store";
+} from "../../../store/onboarding-store";
 import { OnboardingModal } from "./onboarding-modal";
 
 /**
