+++ src/renderer/src/features/onboarding/components/onboarding-modal.tsx
