+++ src/renderer/src/features/onboarding/components/onboarding-styled.tsx
