+++ src/renderer/src/features/onboarding/components/steps/create-agent-step.tsx
