+++ src/renderer/src/features/onboarding/components/steps/default-model-step.tsx
