+++ src/renderer/src/features/onboarding/components/steps/model-credential-step.tsx
