+++ src/renderer/src/features/onboarding/components/steps/radient-choice-step.tsx
