+++ src/renderer/src/features/onboarding/components/steps/radient-signin-step.tsx
