+++ src/renderer/src/features/onboarding/components/steps/search-api-step.tsx
