+++ src/renderer/src/features/onboarding/components/steps/user-profile-step.tsx
