+++ src/renderer/src/features/onboarding/components/steps/welcome-step.tsx
